backend/requirements.txt
Flask==3.1.1
python-dotenv==1.1.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1 
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
flask-cors==6.0.1
python-dateutil
pandas==2.3.0
xlsxwriter==3.2.5
gunicorn==23.0.0